+++ src/main/java/com/Ultra_Nerd/CodeLyokoLegacy/Util/client/sky/volcano/CustomVolcanoSky.java
@@ -34,7 +34,7 @@ public record CustomVolcanoSky() implements DimensionRenderingRegistry.SkyRender
 
         final MatrixStack matrixStack = ctx.matrixStack();
 
-        for (byte i = 0; i < 6; ++i) {
+        for (int i = 0; i < 6; ++i) {
             matrixStack.push();
             switch (i) {
                 case 1 -> {
